# test(eslint-plugin): split up naming-convention