--- a/packages/eslint-plugin/tests/rules/naming-convention/cases/enumMember.test.ts
+++ b/packages/eslint-plugin/tests/rules/naming-convention/cases/enumMember.test.ts
@@ -1,0 +1,10 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: ['enum Ignored { % }', 'enum Ignored { "%" }'],
+    options: {
+      selector: 'enumMember',
+    },
+  },
+]);