--- a/packages/eslint-plugin/tests/rules/naming-convention/cases/function.test.ts
+++ b/packages/eslint-plugin/tests/rules/naming-convention/cases/function.test.ts
@@ -1,0 +1,10 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: ['function % () {}', '(function % () {});', 'declare function % ();'],
+    options: {
+      selector: 'function',
+    },
+  },
+]);