--- a/packages/eslint-plugin/tests/rules/naming-convention/cases/parameterProperty.test.ts
+++ b/packages/eslint-plugin/tests/rules/naming-convention/cases/parameterProperty.test.ts
@@ -1,0 +1,21 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: [
+      'class Ignored { constructor(private %) {} }',
+      'class Ignored { constructor(readonly %) {} }',
+      'class Ignored { constructor(private readonly %) {} }',
+    ],
+    options: {
+      selector: 'parameterProperty',
+    },
+  },
+  {
+    code: ['class Ignored { constructor(private readonly %) {} }'],
+    options: {
+      selector: 'parameterProperty',
+      modifiers: ['readonly'],
+    },
+  },
+]);