--- a/packages/eslint-plugin/tests/rules/naming-convention/cases/typeParameter.test.ts
+++ b/packages/eslint-plugin/tests/rules/naming-convention/cases/typeParameter.test.ts
@@ -1,0 +1,15 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: [
+      'class Ignored<%> {}',
+      'function ignored<%>() {}',
+      'type Ignored<%> = { ignored: % };',
+      'interface Ignored<%> extends Ignored<string> {}',
+    ],
+    options: {
+      selector: 'typeParameter',
+    },
+  },
+]);