--- a/packages/eslint-plugin/tests/rules/naming-convention/cases/variable.test.ts
+++ b/packages/eslint-plugin/tests/rules/naming-convention/cases/variable.test.ts
@@ -1,0 +1,20 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: [
+      'const % = 1;',
+      'let % = 1;',
+      'var % = 1;',
+      'const {%} = {ignored: 1};',
+      'const {% = 2} = {ignored: 1};',
+      'const {...%} = {ignored: 1};',
+      'const [%] = [1];',
+      'const [% = 1] = [1];',
+      'const [...%] = [1];',
+    ],
+    options: {
+      selector: 'variable',
+    },
+  },
+]);