… packages/eslint-plugin/tests/rules/naming-convention/naming-convention.test.ts