## PR Checklist

-   [x] Addresses an existing issue: fixes #4598
-   [x] That issue was marked as [accepting prs](https://github.com/typescript-eslint/typescript-eslint/issues?q=is%3Aopen+is%3Aissue+label%3A%22accepting+prs%22)
-   [x] Steps in [CONTRIBUTING.md](https://github.com/typescript-eslint/typescript-eslint/blob/main/CONTRIBUTING.md) were taken

## Overview

Splits out the dynamically generated tests from `naming-convention.test.ts` into their own test case files. This lets Jest run them in parallel. There should be no changes to test logic.

### Performance runs

-   Execution: `hyperfine -p 'yarn jest --clearCache' --show-output` within `packages/eslint-plugin`. Terminal was in focus and no actions taken on open programs during test runs.
-   Hardware: Surface Laptop 4, 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz.
-   Software: WSL 2.0 in an Ubuntu drive via a VS Code terminal.

| Command                       | `main`                 | `naming-convention-split` | (old - new) / old x 100% |
| ----------------------------- | ---------------------- | ------------------------- | ------------------------ |
| `yarn jest`                   | 206.237 s ±  3.049 s   | 174.888 s ±  0.958 s      | **15%** Faster           |
| `yarn jest naming-convention` | 120.003 s ±  1.357 s   | 60.726 s ±  0.911 s       | **49%** Faster           |

<details>
<summary>Raw data</summary>

`yarn jest` on `main` _(reusing time from another PR)_:

```
Time (mean ± σ):     206.237 s ±  3.049 s    [User: 1156.814 s, System: 40.173 s]
Range (min … max):   201.867 s … 212.914 s    25 runs
```

`yarn jest` on `naming-convention-split`:

```
Time (mean ± σ):     174.888 s ±  0.958 s    [User: 1290.055 s, System: 47.218 s]
Range (min … max):   173.611 s … 176.374 s    10 runs
```

`yarn jest naming-convention` on `main`:

```
Time (mean ± σ):     120.003 s ±  1.357 s    [User: 133.107 s, System: 2.962 s]
Range (min … max):   118.722 s … 122.866 s    10 runs
```

`yarn jest naming-convention` on `naming-convention-split`:

```
Time (mean ± σ):     60.726 s ±  0.911 s    [User: 436.456 s, System: 13.817 s]
Range (min … max):   59.279 s … 62.716 s    10 runs
```

</summary>